KAN-125 feat: 자동 저장 로직 페이지에서 관리하도록 수정
